comments done

+++ spritesheet.cpp
@@ -1,32 +1,15 @@
-/**
- * @Authors: ....
- *
- * The implementation of the spritesheet class
- */
-
 #include <QCoreApplication>
 #include <QJsonDocument>
 #include <QFile>
 #include "spritesheet.h"
-#include <iostream>
 
-/**
- * Constructs the a QObject and current status of a palette
- * @param parent - pointer to a parent QObject
- */
 Spritesheet::Spritesheet(QObject *parent)
     : QObject{parent}, currentFrame(0), palette(this)
 {
     QObject::connect(&timeline, &Timeline::goToNextSprite, this, &Spritesheet::goToNextSprite);
 }
 
-/**
- * This method saves the current sprite to a JSON for the user
- * @param filePath - the current file path where the user wants to save
- * @return - true if file was able to be saved, false otherwise
- */
-void Spritesheet::saveToJson(QString &filePath)
-{
+void Spritesheet::saveToJson(QString &filePath) {
     QJsonDocument jsonDoc(toJson());
     QByteArray jsonData = jsonDoc.toJson(QJsonDocument::Indented);
 
@@ -40,13 +23,7 @@ void Spritesheet::saveToJson(QString &filePath)
     }
 }
 
-/**
- * This method loads a sprite from a JSON to the current sprite sheet
- * @param filePath - the current file path the user is using to load sprite
- * @return - true if the file was able to be loaded, false otherwise
- */
-void Spritesheet::loadJson(QString &filePath)
-{
+void Spritesheet::loadJson(QString &filePath) {
 
     // Check the file at the specified path
     QFile file(filePath);
@@ -97,13 +74,7 @@ void Spritesheet::loadJson(QString &filePath)
     emitIfNextOrPreviousSprites();
 }
 
-/**
- * This method creates a new instance of a frame/sprite for the user to work on
- * @param newWidth - the new width value
- * @param newHeight - the new height value
- */
-void Spritesheet::newProject(int newWidth, int newHeight)
-{
+void Spritesheet::newProject(int newWidth, int newHeight) {
     sprites.clear();
 
     width = newWidth;
@@ -121,61 +92,35 @@ void Spritesheet::newProject(int newWidth, int newHeight)
     emitIfNextOrPreviousSprites();
 }
 
-/**
- * This method returns the current sprite from the spritesheet
- * @return - the current sprite
- */
-Sprite &Spritesheet::getCurrentSprite()
-{
+Sprite &Spritesheet::getCurrentSprite() {
     return sprites[currentFrame];
 }
 
-/**
- * This method gets the current coordinates of where the
- * current sprite is being painted
- * @param x - the x coordinate
- * @param y - the y coordinate
- */
-void Spritesheet::paintedCurrentSpriteAt(int x, int y)
-{
+void Spritesheet::paintedCurrentSpriteAt(int x, int y) {
     int brushRadius = palette.getBrush()->getRadius();
     getCurrentSprite().paintAt(x, y, brushRadius, palette.getCurrentColor());
 }
 
-/**
- * This method gets the current coordinates of where the current
- * sprite is erased
- * @param x - the x coordinate
- * @param y - the y coordinate
- */
-void Spritesheet::erasedCurrentSpriteAt(int x, int y)
-{
+void Spritesheet::erasedCurrentSpriteAt(int x, int y) {
     int brushRadius = palette.getBrush()->getRadius();
     getCurrentSprite().eraseAt(x, y, brushRadius);
 }
 
-/**
- * This method returns the current palette
- * @return - the current palette
- */
-Palette &Spritesheet::getPalette()
-{
+Palette &Spritesheet::getPalette() {
     return palette;
 }
 
-Timeline &Spritesheet::getTimeline()
-{
+Timeline &Spritesheet::getTimeline() {
     return timeline;
 }
 
-void Spritesheet::goToNextSprite()
-{
+void Spritesheet::goToNextSprite() {
     if (sprites.size() <= 1)
         return;
 
     currentFrame++;
 
-    if (currentFrame >= sprites.size())
+    if (currentFrame >= (int)sprites.size())
     {
         currentFrame = 0;
     }
@@ -185,8 +130,7 @@ void Spritesheet::goToNextSprite()
     emitIfNextOrPreviousSprites();
 }
 
-void Spritesheet::goToPreviousSprite()
-{
+void Spritesheet::goToPreviousSprite() {
     if (sprites.size() <= 1)
         return;
 
@@ -202,8 +146,7 @@ void Spritesheet::goToPreviousSprite()
     emitIfNextOrPreviousSprites();
 }
 
-void Spritesheet::addSprite()
-{
+void Spritesheet::addSprite() {
     sprites.push_back(Sprite(width, height));
 
     currentFrame = sprites.size() - 1;
@@ -219,8 +162,7 @@ void Spritesheet::addSprite()
     emit currentSpriteID(currentFrame);
 }
 
-void Spritesheet::removeSprite()
-{
+void Spritesheet::removeSprite() {
     sprites.erase(sprites.begin() + currentFrame);
 
     if (sprites.size() <= 1)
@@ -243,11 +185,12 @@ void Spritesheet::removeSprite()
     emit currentSpriteID(currentFrame);
 }
 
-void Spritesheet::emitIfNextOrPreviousSprites()
-{
+void Spritesheet::emitIfNextOrPreviousSprites() {
     bool hasPrevious = (currentFrame > 0) && !sprites.empty();
-    bool hasNext = (currentFrame < sprites.size() - 1) && !sprites.empty();
+    bool hasNext = (currentFrame < (int)sprites.size() - 1) && !sprites.empty();
 
     emit isTherePreviousSprite(hasPrevious);
     emit isThereNextSprite(hasNext);
 }
+
+// - Checked by Roland Guerriere
